services/requirements.txt
absl-py==2.3.0
annotated-types==0.7.0
anyio==4.9.0
astunparse==1.6.3
attrs==25.4.0
audioread==3.0.1
certifi==2025.6.15
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
cmake==4.0.3
contourpy==1.3.3
cycler==0.12.1
decorator==5.2.1
fastapi==0.115.13
filelock==3.18.0
flatbuffers==25.9.23
fonttools==4.61.1
fsspec==2025.5.1
gast==0.6.0
google-pasta==0.2.0
grpcio==1.73.1
h11==0.16.0
h5py==3.14.0
hf-xet==1.1.5
huggingface-hub==0.33.2
idna==3.10
img2vec_pytorch==1.0.2
Jinja2==3.1.6
joblib==1.5.1
keras==3.10.0
kiwisolver==1.4.9
lazy_loader==0.4
libclang==18.1.1
librosa==0.11.0
llvmlite==0.44.0
Markdown==3.8.2
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.8
mdurl==0.1.2
mediapipe==0.10.7
ml-dtypes==0.4.1
mpmath==1.3.0
msgpack==1.1.1
namex==0.1.0
networkx==3.5
numba==0.61.2
numpy==2.0.2
opencv-contrib-python==4.12.0.88
opencv-python==4.12.0.88
opt_einsum==3.4.0
optree==0.16.0
packaging==25.0
pillow==11.2.1
platformdirs==4.3.8
pooch==1.8.2
protobuf==3.20.3
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
Pygments==2.19.2
pyparsing==3.2.5
python-dateutil==2.9.0.post0
python-multipart==0.0.20
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.4
rich==14.0.0
safetensors==0.5.3
scikit-learn==1.7.0
scipy==1.15.3
six==1.17.0
sniffio==1.3.1
sounddevice==0.5.3
soundfile==0.13.1
soxr==0.5.0.post1
SQLAlchemy==2.0.45
starlette==0.46.2
sympy==1.14.0
tensorboard==2.18.0
tensorboard-data-server==0.7.2
tensorflow==2.18.0
tensorflow-io-gcs-filesystem==0.37.1
termcolor==3.1.0
threadpoolctl==3.6.0
tokenizers==0.21.2
torch==2.9.1
torchvision==0.24.1
tqdm==4.67.1
transformers==4.53.1
typing-inspection==0.4.1
typing_extensions==4.14.0
urllib3==2.5.0
uvicorn==0.34.3
Werkzeug==3.1.3
wrapt==1.17.2
